No need to inject on runtime a workplace search client
